+++ CHANGELOG.md
@@ -32,6 +32,7 @@ BUG FIXES:
 
 IMPROVEMENTS:
 - [rpc] `/tx` and `/tx_search` responses now include the transaction hash
+- [rpc] include validator power in `/status`
 
 ## 0.17.1 (March 27th, 2018)